refactor: 예외 핸들링은 일단 한 클래스에서 진행
